pyproject.toml
[project]
name = "lab-tic-tac-toe-step-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.0",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
    "uvicorn>=0.35.0",
]
